+++ frontend/app/.server/domain/services/city-service-default.ts
@@ -2,8 +2,8 @@ import { Ok, Err } from 'oxide.ts';
 import type { Result, Option } from 'oxide.ts';
 
 import type { City, LocalizedCity } from '~/.server/domain/models';
+import { apiFetch } from '~/.server/domain/services/api-client';
 import type { CityService } from '~/.server/domain/services/city-service';
-import { apiFetch } from '~/.server/domain/services/make-api-request';
 import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 import { HttpStatusCodes } from '~/errors/http-status-codes';
