+++ frontend/app/.server/domain/services/directorate-service-default.ts
@@ -2,8 +2,8 @@ import { Ok, Err } from 'oxide.ts';
 import type { Result, Option } from 'oxide.ts';
 
 import type { LocalizedDirectorate, Directorate, Branch } from '~/.server/domain/models';
+import { apiFetch } from '~/.server/domain/services/api-client';
 import type { DirectorateService } from '~/.server/domain/services/directorate-service';
-import { apiFetch } from '~/.server/domain/services/make-api-request';
 import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 import { HttpStatusCodes } from '~/errors/http-status-codes';
